fix: resolve all 138 pipeline mypy errors (#228)
* fix: resolve all 138 pipeline mypy errors

- Add per-module mypy override to properly suppress ai_engine/model_vehicle
  (39 errors — exclude pattern doesn't work for followed imports)
- Add types-PyYAML dev dependency
- Suppress mixin self-type errors via disable_error_code for health,
  message_ops, kafka_setup modules (known mypy mixin limitation)
- Add type annotations to service_base, transformer, fiber_config,
  and processing steps
- Fix base_step.process return type to allow None (matches all subclasses)
- Fix bandpass filter to return x instead of y (same value, always defined)
- Guard None comparisons in transformer/ai_engine rolling buffers
- Widen gauge parameter from int to float in model_T (was already float
  in config, truncation would change speed calculations)

Pipeline mypy: 138 errors → 0. All changes are type annotations,
assert guards, and mypy config — zero runtime behavior changes.

Closes #60

* fix: harden None-narrowing guards across pipeline and backend

- message_ops: use `is not None` instead of truthiness check to prevent
  _pending_deliveries memory leak for empty-string message IDs
- service_base: correct _pending_deliveries type to dict[str, Any]
- fiber_config: replace assert with explicit RuntimeError (survives -O)
- detection_api, export_views: add None guards after parse/select_tier
- simulation: add type annotation for ring variable

Resolves all 24 pre-existing backend mypy errors.

+++ services/pipeline/ai_engine/model_vehicle/model_T.py
@@ -41,7 +41,7 @@ class Args_NN_model_all_channels:
     def __init__(
         self,
         data_window_length: int,
-        gauge: int,
+        gauge: float,
         Nch: int,
         N_channels: int,
         fs: float,
